Inserted short line to deal with preventable error.
For some reason, Scriptogram's template puts its braced variables
into the CSS right up against closing braces. Although Scriptogram
is able to cope with this, testing shows that vanilla Mustache is
not.

This line simply separates the double brace at the end of Mustache
variable from the closing brace of a CSS property and avoids an
error.