Added code which will allow proper HTML rendering.
Somehow, Scriptogram's template replaces a double-braced `content`
variable with functioning HTML rather than encoded HTML which is
what Mustache double-braced variables normally give you.

Rather than change the variables to triple-braces directly in the
template, only to have to potentially change them back again
for use in Scriptogram, this added code will check for an array
of variables to match to and replace double-braced instaces with
triple-braced instances.

This ensures that posts display as fully rendered Markdown in
the preview and not encoded HTML.