Removed specific context values from main script and added `setup.js`.
Running `setup.js` creates an empty `context.json` file so that
this can be filled in with custom values relevant to whoever is
using it. It takes my blog values out of the main script but
also means no-one else has to go into the main script to tinker.

Some basic error-handling has been added.