Based tag display on whether or not tags exist.
Had automatically set the `if_tags` of each post object to `true`
regardless of what tags might be available. Have now based this
property on the length of the array, meaning it will return as
falsy if there are no tags at all to be found. This will stop
Mustache from rendering this part of the template.