Added opt-in function for `ahem` workflow.
Whereas previously the `ahem` workflow was an assumption, now
one must declare it as a true property in `config.json` to run
it this way.

The default is now that the user will want to preview Scriptogram
as if using it in a vanilla way, with all posts in one place and
with all metadata in the files themselves.

As such the property `ahem` is added to the config setup by
`setup.js` but it is automatically set to `false`.

The console should log which mode the user is in when the preview
is run.